+++ providers/sdk/java/NOTICE
