+++ providers/sdk/java/docs/conf.py
@@ -1,3 +1,5 @@
+# Disable Flake8 because of all the sphinx imports
+#
 # Licensed to the Apache Software Foundation (ASF) under one
 # or more contributor license agreements.  See the NOTICE file
 # distributed with this work for additional information
@@ -14,26 +16,12 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-#
-# NOTE! THIS FILE IS AUTOMATICALLY GENERATED AND WILL BE
-# OVERWRITTEN WHEN PREPARING DOCUMENTATION FOR THE PACKAGES.
-#
-# IF YOU WANT TO MODIFY THIS FILE, YOU SHOULD MODIFY THE TEMPLATE
-# `PROVIDER__INIT__PY_TEMPLATE.py.jinja2` IN the `dev/breeze/src/airflow_breeze/templates` DIRECTORY
-#
+"""Configuration of Providers docs building."""
+
 from __future__ import annotations
 
-import packaging.version
+import os
 
-from airflow import __version__ as airflow_version
+os.environ["AIRFLOW_PACKAGE_NAME"] = "apache-airflow-providers-sdk-java"
 
-__all__ = ["__version__"]
-
-__version__ = "0.1.0"
-
-if packaging.version.parse(packaging.version.parse(airflow_version).base_version) < packaging.version.parse(
-    "3.0.0"
-):
-    raise RuntimeError(
-        f"The package `apache-airflow-providers-languages-java:{__version__}` needs Apache Airflow 3.0.0+"
-    )
+from docs.provider_conf import *  # noqa: F403
