+++ providers/sdk/java/pyproject.toml
@@ -24,9 +24,9 @@ requires = ["flit_core==3.12.0"]
 build-backend = "flit_core.buildapi"
 
 [project]
-name = "apache-airflow-providers-languages-java"
+name = "apache-airflow-providers-sdk-java"
 version = "0.1.0"
-description = "Provider package apache-airflow-providers-languages-java for Apache Airflow"
+description = "Provider package apache-airflow-providers-sdk-java for Apache Airflow"
 readme = "README.rst"
 license = "Apache-2.0"
 license-files = ['LICENSE', 'NOTICE']
@@ -36,7 +36,7 @@ authors = [
 maintainers = [
     {name="Apache Software Foundation", email="dev@airflow.apache.org"},
 ]
-keywords = [ "airflow-provider", "languages.java", "airflow", "integration" ]
+keywords = [ "airflow-provider", "sdk.java", "airflow", "integration" ]
 classifiers = [
     "Development Status :: 5 - Production/Stable",
     "Environment :: Console",
@@ -104,8 +104,8 @@ apache-airflow-providers-common-sql = {workspace = true}
 apache-airflow-providers-standard = {workspace = true}
 
 [project.urls]
-"Documentation" = "https://airflow.apache.org/docs/apache-airflow-providers-languages-java/0.1.0"
-"Changelog" = "https://airflow.apache.org/docs/apache-airflow-providers-languages-java/0.1.0/changelog.html"
+"Documentation" = "https://airflow.apache.org/docs/apache-airflow-providers-sdk-java/0.1.0"
+"Changelog" = "https://airflow.apache.org/docs/apache-airflow-providers-sdk-java/0.1.0/changelog.html"
 "Bug Tracker" = "https://github.com/apache/airflow/issues"
 "Source Code" = "https://github.com/apache/airflow"
 "Slack Chat" = "https://s.apache.org/airflow-slack"
@@ -113,7 +113,7 @@ apache-airflow-providers-standard = {workspace = true}
 "YouTube" = "https://www.youtube.com/channel/UCSXwxpWZQ7XZ1WL3wqevChA/"
 
 [project.entry-points."apache_airflow_provider"]
-provider_info = "airflow.providers.languages.java.get_provider_info:get_provider_info"
+provider_info = "airflow.providers.sdk.java.get_provider_info:get_provider_info"
 
 [tool.flit.module]
-name = "airflow.providers.languages.java"
+name = "airflow.providers.sdk.java"
