+++ pyproject.toml
@@ -266,9 +266,6 @@ apache-airflow = "airflow.__main__:main"
 "keycloak" = [
     "apache-airflow-providers-keycloak>=0.0.1"
 ]
-"languages.java" = [
-    "apache-airflow-providers-languages-java>=0.1.0" # Set from local provider pyproject.toml
-]
 "microsoft.azure" = [
     "apache-airflow-providers-microsoft-azure>=10.5.1"
 ]
@@ -340,6 +337,9 @@ apache-airflow = "airflow.__main__:main"
 ]
 "samba" = [
     "apache-airflow-providers-samba>=4.9.0"
+]
+"sdk.java" = [
+    "apache-airflow-providers-sdk-java>=0.1.0" # Set from local provider pyproject.toml
 ]
 "segment" = [
     "apache-airflow-providers-segment>=3.7.0"
@@ -460,7 +460,6 @@ apache-airflow = "airflow.__main__:main"
     "apache-airflow-providers-jdbc>=4.5.2",
     "apache-airflow-providers-jenkins>=3.7.2",
     "apache-airflow-providers-keycloak>=0.0.1",
-    "apache-airflow-providers-languages-java>=0.1.0", # Set from local provider pyproject.toml
     "apache-airflow-providers-microsoft-azure>=10.5.1",
     "apache-airflow-providers-microsoft-mssql>=3.9.2",
     "apache-airflow-providers-microsoft-psrp>=3.0.0",
@@ -485,6 +484,7 @@ apache-airflow = "airflow.__main__:main"
     "apache-airflow-providers-redis>=4.0.0",
     "apache-airflow-providers-salesforce>=5.9.0",
     "apache-airflow-providers-samba>=4.9.0",
+    "apache-airflow-providers-sdk-java>=0.1.0", # Set from local provider pyproject.toml
     "apache-airflow-providers-segment>=3.7.0",
     "apache-airflow-providers-sendgrid>=4.0.0",
     "apache-airflow-providers-sftp>=5.0.0",
@@ -1174,8 +1174,6 @@ mypy_path = [
     "$MYPY_CONFIG_FILE_DIR/providers/jenkins/tests",
     "$MYPY_CONFIG_FILE_DIR/providers/keycloak/src",
     "$MYPY_CONFIG_FILE_DIR/providers/keycloak/tests",
-    "$MYPY_CONFIG_FILE_DIR/providers/languages/java/src",
-    "$MYPY_CONFIG_FILE_DIR/providers/languages/java/tests",
     "$MYPY_CONFIG_FILE_DIR/providers/microsoft/azure/src",
     "$MYPY_CONFIG_FILE_DIR/providers/microsoft/azure/tests",
     "$MYPY_CONFIG_FILE_DIR/providers/microsoft/mssql/src",
@@ -1224,6 +1222,8 @@ mypy_path = [
     "$MYPY_CONFIG_FILE_DIR/providers/salesforce/tests",
     "$MYPY_CONFIG_FILE_DIR/providers/samba/src",
     "$MYPY_CONFIG_FILE_DIR/providers/samba/tests",
+    "$MYPY_CONFIG_FILE_DIR/providers/sdk/java/src",
+    "$MYPY_CONFIG_FILE_DIR/providers/sdk/java/tests",
     "$MYPY_CONFIG_FILE_DIR/providers/segment/src",
     "$MYPY_CONFIG_FILE_DIR/providers/segment/tests",
     "$MYPY_CONFIG_FILE_DIR/providers/sendgrid/src",
@@ -1443,7 +1443,6 @@ apache-airflow-providers-informatica = false
 apache-airflow-providers-jdbc = false
 apache-airflow-providers-jenkins = false
 apache-airflow-providers-keycloak = false
-apache-airflow-providers-languages-java = false
 apache-airflow-providers-microsoft-azure = false
 apache-airflow-providers-microsoft-mssql = false
 apache-airflow-providers-microsoft-psrp = false
@@ -1468,6 +1467,7 @@ apache-airflow-providers-qdrant = false
 apache-airflow-providers-redis = false
 apache-airflow-providers-salesforce = false
 apache-airflow-providers-samba = false
+apache-airflow-providers-sdk-java = false
 apache-airflow-providers-segment = false
 apache-airflow-providers-sendgrid = false
 apache-airflow-providers-sftp = false
@@ -1595,7 +1595,6 @@ apache-airflow-providers-informatica = false
 apache-airflow-providers-jdbc = false
 apache-airflow-providers-jenkins = false
 apache-airflow-providers-keycloak = false
-apache-airflow-providers-languages-java = false
 apache-airflow-providers-microsoft-azure = false
 apache-airflow-providers-microsoft-mssql = false
 apache-airflow-providers-microsoft-psrp = false
@@ -1620,6 +1619,7 @@ apache-airflow-providers-qdrant = false
 apache-airflow-providers-redis = false
 apache-airflow-providers-salesforce = false
 apache-airflow-providers-samba = false
+apache-airflow-providers-sdk-java = false
 apache-airflow-providers-segment = false
 apache-airflow-providers-sendgrid = false
 apache-airflow-providers-sftp = false
@@ -1757,7 +1757,6 @@ apache-airflow-providers-informatica = { workspace = true }
 apache-airflow-providers-jdbc = { workspace = true }
 apache-airflow-providers-jenkins = { workspace = true }
 apache-airflow-providers-keycloak = { workspace = true }
-apache-airflow-providers-languages-java = { workspace = true }
 apache-airflow-providers-microsoft-azure = { workspace = true }
 apache-airflow-providers-microsoft-mssql = { workspace = true }
 apache-airflow-providers-microsoft-psrp = { workspace = true }
@@ -1782,6 +1781,7 @@ apache-airflow-providers-qdrant = { workspace = true }
 apache-airflow-providers-redis = { workspace = true }
 apache-airflow-providers-salesforce = { workspace = true }
 apache-airflow-providers-samba = { workspace = true }
+apache-airflow-providers-sdk-java = { workspace = true }
 apache-airflow-providers-segment = { workspace = true }
 apache-airflow-providers-sendgrid = { workspace = true }
 apache-airflow-providers-sftp = { workspace = true }
@@ -1895,7 +1895,6 @@ members = [
     "providers/jdbc",
     "providers/jenkins",
     "providers/keycloak",
-    "providers/languages/java",
     "providers/microsoft/azure",
     "providers/microsoft/mssql",
     "providers/microsoft/psrp",
@@ -1920,6 +1919,7 @@ members = [
     "providers/redis",
     "providers/salesforce",
     "providers/samba",
+    "providers/sdk/java",
     "providers/segment",
     "providers/sendgrid",
     "providers/sftp",
