+++ scripts/ci/docker-compose/remove-sources.yml
@@ -83,7 +83,6 @@ services:
       - ../../../empty:/opt/airflow/providers/jdbc/src
       - ../../../empty:/opt/airflow/providers/jenkins/src
       - ../../../empty:/opt/airflow/providers/keycloak/src
-      - ../../../empty:/opt/airflow/providers/languages/java/src
       - ../../../empty:/opt/airflow/providers/microsoft/azure/src
       - ../../../empty:/opt/airflow/providers/microsoft/mssql/src
       - ../../../empty:/opt/airflow/providers/microsoft/psrp/src
@@ -108,6 +107,7 @@ services:
       - ../../../empty:/opt/airflow/providers/redis/src
       - ../../../empty:/opt/airflow/providers/salesforce/src
       - ../../../empty:/opt/airflow/providers/samba/src
+      - ../../../empty:/opt/airflow/providers/sdk/java/src
       - ../../../empty:/opt/airflow/providers/segment/src
       - ../../../empty:/opt/airflow/providers/sendgrid/src
       - ../../../empty:/opt/airflow/providers/sftp/src
