+++ scripts/ci/docker-compose/tests-sources.yml
@@ -96,7 +96,6 @@ services:
       - ../../../providers/jdbc/tests:/opt/airflow/providers/jdbc/tests
       - ../../../providers/jenkins/tests:/opt/airflow/providers/jenkins/tests
       - ../../../providers/keycloak/tests:/opt/airflow/providers/keycloak/tests
-      - ../../../providers/languages/java/tests:/opt/airflow/providers/languages/java/tests
       - ../../../providers/microsoft/azure/tests:/opt/airflow/providers/microsoft/azure/tests
       - ../../../providers/microsoft/mssql/tests:/opt/airflow/providers/microsoft/mssql/tests
       - ../../../providers/microsoft/psrp/tests:/opt/airflow/providers/microsoft/psrp/tests
@@ -121,6 +120,7 @@ services:
       - ../../../providers/redis/tests:/opt/airflow/providers/redis/tests
       - ../../../providers/salesforce/tests:/opt/airflow/providers/salesforce/tests
       - ../../../providers/samba/tests:/opt/airflow/providers/samba/tests
+      - ../../../providers/sdk/java/tests:/opt/airflow/providers/sdk/java/tests
       - ../../../providers/segment/tests:/opt/airflow/providers/segment/tests
       - ../../../providers/sendgrid/tests:/opt/airflow/providers/sendgrid/tests
       - ../../../providers/sftp/tests:/opt/airflow/providers/sftp/tests
